fix patch strip
